refactor: 코드 개선 - Lotto 는 단순히 로또 값을 나타내는 vo 이므로 vo 패키지로 이동 - 기존 Game 클래스를 결과를 판단하는 GameResultJudge 클래스와 수익률을 계산하는 GameReturnRateCalculator 로 분리함
