+++ src/main/java/lotto/Application.java
@@ -1,12 +1,11 @@
 package lotto;
 
 import lotto.controller.GameController;
-import lotto.domain.Game;
 
 public class Application {
 
     public static void main(String[] args) {
-        GameController controller = new GameController(new Game());
+        GameController controller = new GameController();
         controller.gameStart();
     }
 }
