+++ src/main/java/lotto/vo/Lotto.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.vo;
 
 import java.util.List;
 import lotto.validator.LottoValidator;
